requirements.txt
python-digitalocean==1.13.2
click==6.6
